move non-templated code to .cpp files
It is standard practice to put non-templated code into .cpp files. This has multiple benefits, including reducing compile times and less leackage of private libraries into the public interfaces.
